register controller and postman

+++ src/utils/cloudinary.js
@@ -1,5 +1,5 @@
-import {v2 as cloudinary} from "cloudinary" 
-import fs from "fs"
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -7,22 +7,27 @@ cloudinary.config({
     api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
-
-const uploadOnCloudinary  = async (localFilePath) => {
+const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if (!localFilePath) return null
-      //upload the file on cloudinary
-       const response = await cloudinary.uploader.upload(localFilePath,{
-            resource_type: "auto"
-        })
-        //file has been uploaded successfull
-        console.log("file is uploaded on cloudinary", response.url);
-        return response;
-    } catch(error){
-        fs.unlinkSync(localFilePath)
-        return null;
-    
-    }
-}
+        if (!localFilePath) return null;
 
-export {cloudinary}
+        const response = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "auto"
+        });
+
+        fs.unlinkSync(localFilePath);
+
+        return response;
+
+    } catch (error) {
+        console.log("Cloudinary error:", error);
+
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
+        return null;
+    }
+};
+
+export { uploadOnCloudinary };
